aws-lambda/clim-forecast_updates-aws-lambda-docker/climate_image/requirements.txt
boto3
pandas
geopandas
requests
pyjanitor
Shapely
Fiona
pyproj
rtree